requirements.txt
python-docx>=1.1.0
llama-index-core>=0.12.0
llama-index-vector-stores-postgres>=0.4.0
llama-index-embeddings-huggingface>=0.5.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
jieba>=0.42.1
rank-bm25>=0.2.2
FlagEmbedding>=1.2.0
python-dotenv>=1.0.0
vllm
openai>=1.0.0
bitsandbytes>=0.45.0
# ── LangGraph Agent ──
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langmem>=0.0.30
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.1.0
# ── API Server ──
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
# ── MCP & Tools ──
fastmcp>=2.0.0
httpx>=0.27.0
